serve SPSP and verify requests on same port
Resolves #22
