+++ README.md
@@ -11,10 +11,6 @@ The **Receipt Verifier**:
 1. pre-shares a secret key with the receiving wallet for generating receipts, by acting as a proxy for SPSP queries to the recipient's payment pointer
 2. verifies receipts
 
-SPSP queries are proxied to the URL encoded [payment pointer](https://paymentpointers.org/) or [SPSP endpoint](https://interledger.org/rfcs/0009-simple-payment-setup-protocol/) in the path of the SPSP request URL (unless [`SPSP_ENDPOINTS_URL`](#spsp_endpoints_url) is configured).
-
-* Example: An SPSP query to `https://my-receipt-verifier.com/%24wallet.com%2Falice` (or `https://my-receipt-verifier.com/https%3A%2F%2Fwallet.com%2Falice`) is proxied by the **Receipt Verifier** to `$wallet.com/alice` (`https://wallet.com/alice`).
-
 For [Web Monetization](https://github.com/interledger/rfcs/blob/master/0028-web-monetization/0028-web-monetization.md), website visitors submit receipts to the website in `monetizationprogress` events. The website backend can send receipts to the **Receipt Verifier** to confirm the payment.
 
 ### Run
@@ -27,6 +23,11 @@ npm start
 ```
 
 ### Environment Variables
+
+#### PORT
+* Type: [Number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)
+* Description: The port that Receipt Verifier API will listen on.
+* Default: 3000
 
 #### RECEIPT_SEED
 * Type: String
@@ -49,22 +50,19 @@ npm start
 For each SPSP query, a GET request is sent to `SPSP_ENDPOINTS_URL` with the query's url path value (without the preceding slash) as the URI encoded `id` query parameter.
 The response body is expected to be a string of the SPSP endpoint to proxy the SPSP query to.
 
-* Example: if `SPSP_ENDPOINTS_URL=https://my-revshare.com`, an SPSP query to `https://my-receipt-verifier.com/users/alice` triggers a GET request to `https://my-revshare.com/?id=users%2Falice`. The receipt verifier proxies the SPSP query to the `spspEndpoint` in the response.
-
-#### SPSP_PROXY_PORT
-* Type: [Number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)
-* Description: The port that SPSP proxy will listen on.
-* Default: 3001
-
-#### VERIFIER_PORT
-* Type: [Number](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number)
-* Description: The port that Receipt Verifier API will listen on.
-* Default: 3000
-
 ### API Documentation
 
+#### `GET /  Accept: application/spsp4+json`
+Adds receipt headers to [SPSP request](https://interledger.org/rfcs/0009-simple-payment-setup-protocol/) and proxies it to the receiver's SPSP endpoint.
+
+If [`SPSP_ENDPOINTS_URL`](#spsp_endpoints_url) is configured, the request is proxied to the SPSP endpoint returned by the `SPSP_ENDPOINTS_URL`.
+* Example: if `SPSP_ENDPOINTS_URL=https://my-revshare.com`, `GET /users/alice` triggers a GET request to `https://my-revshare.com/?id=users%2Falice`. The SPSP request is then proxied to the SPSP endpoint in the response.
+
+Otherwise, the SPSP query is proxied to the URL encoded [payment pointer](https://paymentpointers.org/) or SPSP endpoint in the path of the SPSP request URL.
+* Example: `GET /%24wallet.com` (or `GET /https%3A%2F%2Fwallet.com%2F.well-known%2Fpay`) is proxied to `$wallet.com` (`https://wallet.com/.well-known/pay`).
+
 #### `POST /receipts`
-Verifies receipt and returns decoded values
+Verifies receipt and returns value
 
 ##### Request Body:
 * Type: String
