+++ src/services/App.ts
@@ -1,28 +1,23 @@
 import { Injector } from 'reduct'
-import { Receipts } from './Receipts'
 import { Redis } from './Redis'
-import { SPSP } from './SPSP'
+import { Server } from './Server'
 
 export class App {
-  private receipts: Receipts
   private redis: Redis
-  private spsp: SPSP
+  private server: Server
 
   constructor (deps: Injector) {
-    this.receipts = deps(Receipts)
     this.redis = deps(Redis)
-    this.spsp = deps(SPSP)
+    this.server = deps(Server)
   }
 
   start (): void {
     this.redis.start()
-    this.receipts.start()
-    this.spsp.start()
+    this.server.start()
   }
 
   async stop (): Promise<void> {
-    this.receipts.stop()
-    this.spsp.stop()
     await this.redis.stop()
+    await this.server.stop()
   }
 }
